finished y-sorting system for the tilemap. Tilemap improvements. More debug options. and other stuff.

+++ tool/map_converter.zig
@@ -36,7 +36,7 @@ const tiled_types = struct {
     };
 
     const TileProperty = struct {
-        value: u8,
+        value: u16,
         name: []const u8,
     };
